github cleanup
