+++ pipelines/github/settings.py
@@ -1,0 +1,8 @@
+START_DATE = "1970-01-01T00:00:00Z"
+
+# rest queries
+REST_API_BASE_URL = "https://api.github.com"
+REPO_EVENTS_PATH = "/repos/%s/%s/events"
+
+# graphql queries
+GRAPHQL_API_BASE_URL = "https://api.github.com/graphql"
