+++ tests/utils.py
@@ -24,9 +24,6 @@ TEST_STORAGE_ROOT = "_storage"
 
 # get env variable with destinations
 ALL_DESTINATIONS = dlt.config.get("ALL_DESTINATIONS", list) or [
-    "bigquery",
-    "redshift",
-    "postgres",
     "duckdb",
 ]
 # ALL_DESTINATIONS = ["duckdb"]
